storage/requirements.txt
greenlet==3.2.3
neo4j==5.19.0
numpy==1.26.4
pandas==2.2.2
psycopg2-binary==2.9.9
pyarrow==15.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.30
typing_extensions==4.14.1
tzdata==2025.2